수정, 삭제부분 완료 갤러리 작업중

+++ community/src/main/java/net/musecom/community/service/BoardService.java
@@ -47,7 +47,7 @@ public class BoardService {
 		return bMapper.setDeleteBoard(id);
 	}
 	
-	public BoardDto getBoardByPass(int id, String password) {
+	public BoardDto getBoardByPass(long id, String password) {
 		Map<String, Object> params = new HashMap<>();
 		params.put("id", id);
 		params.put("password", password);
